Add http session handling to review

+++ src/main/java/webdev/services/ReviewService.java
@@ -7,6 +7,7 @@ import webdev.models.User;
 import webdev.repositories.ReviewRepository;
 import webdev.repositories.UserRepository;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 import java.util.Optional;
 
@@ -44,23 +45,35 @@ public class ReviewService {
     }
 
     @PostMapping("/api/user/{uid}/review")
-    public Review createReview(@PathVariable("uid") int userId, @RequestBody Review review) {
+    public Review createReview(@PathVariable("uid") int userId, @RequestBody Review review, HttpSession httpSession) {
         Optional<User> maybeUser = userRepository.findById(userId);
+        User sessionUser = (User) httpSession.getAttribute("user");
+
         if (maybeUser.isPresent()) {
             User user = maybeUser.get();
-            List<Review> reviews = user.getReviews();
-            reviews.add(review);
-            review.setUser(user);
-            user.setReviews(reviews);
-            userRepository.save(user);
-            reviewRepository.save(review);
-            return review;
+            if (user.getId() == sessionUser.getId()) {
+                List<Review> reviews = user.getReviews();
+                reviews.add(review);
+                review.setUser(user);
+                user.setReviews(reviews);
+                userRepository.save(user);
+                reviewRepository.save(review);
+                return review;
+            }
         }
         return null;
     }
 
     @DeleteMapping("/api/review/{rid}")
-    public void deleteReview(@PathVariable("rid") int reviewId) {
-        reviewRepository.deleteById(reviewId);
+    public void deleteReview(@PathVariable("rid") int reviewId, HttpSession httpSession) {
+        User sessionUser = (User) httpSession.getAttribute("user");
+        Optional<Review> maybeReview = reviewRepository.findById(reviewId);
+
+        if (maybeReview.isPresent()) {
+            Review review = maybeReview.get();
+            if (review.getUser().getId() == sessionUser.getId()) {
+                reviewRepository.deleteById(reviewId);
+            }
+        }
     }
 }
